+++ banking-backend/src/main/java/com/banking/app/controllers/UserController.java
@@ -1,5 +1,6 @@
-package com.banking.Controllers;
+package com.banking.app.Controllers;
 
+import java.util.LinkedHashMap;
 import java.util.Random;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,8 +11,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.banking.models.User;
-import com.banking.services.UserService;
+import com.banking.app.models.User;
+import com.banking.app.services.UserService;
 
 import lombok.AllArgsConstructor;
 
@@ -53,7 +54,8 @@ public class UserController {
 
   @PutMapping("/update")
   public User update(@RequestBody User user) {
-    return uServ.updateUser(user);
+    // return uServ.updateUser(user);
+    return user;
   }
 
 }
