+++ banking-backend/src/main/java/com/banking/app/exceptions/CannotUpdateUserException.java
@@ -1,0 +1,14 @@
+package com.banking.app.exceptions;
+
+public class CannotUpdateUserException extends RuntimeException {
+
+  /**
+   * 
+   */
+  private static final long serialVersionUID = 1L;
+
+  public CannotUpdateUserException() {
+    super("Cannot update user.");
+  }
+
+}
