+++ banking-backend/src/main/java/com/banking/app/exceptions/ClaimAlreadyProcessedException.java
@@ -1,0 +1,14 @@
+package com.banking.app.exceptions;
+
+public class ClaimAlreadyProcessedException extends RuntimeException {
+
+  /**
+   * 
+   */
+  private static final long serialVersionUID = 1L;
+
+  public ClaimAlreadyProcessedException() {
+    super("Claim already processed.");
+  }
+
+}
