+++ banking-backend/src/main/java/com/banking/app/exceptions/InvalidCredentialsException.java
@@ -1,0 +1,14 @@
+package com.banking.app.exceptions;
+
+public class InvalidCredentialsException extends RuntimeException {
+
+  /**
+   * 
+   */
+  private static final long serialVersionUID = 1L;
+
+  public InvalidCredentialsException() {
+    super("Incorrect username or password.");
+  }
+
+}
