+++ banking-backend/src/main/java/com/banking/app/models/AccountType.java
@@ -1,4 +1,4 @@
-package com.banking.models;
+package com.banking.app.models;
 
 public enum AccountType {
   CHECKING, SAVINGS
