+++ banking-backend/src/main/java/com/banking/app/models/TransactionData.java
@@ -1,4 +1,4 @@
-package com.banking.models;
+package com.banking.app.models;
 
 import java.time.LocalDate;
