+++ banking-backend/src/main/java/com/banking/app/models/User.java
@@ -1,8 +1,8 @@
-package com.banking.models;
+package com.banking.app.models;
 
 import java.util.List;
 
-import com.banking.utils.IdGenerator;
+import com.banking.app.utils.IdGenerator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
