+++ banking-backend/src/main/java/com/banking/app/models/UserType.java
@@ -1,4 +1,4 @@
-package com.banking.models;
+package com.banking.app.models;
 
 public enum UserType {
