+++ banking-backend/src/main/java/com/banking/app/repositories/UserRepository.java
@@ -1,0 +1,13 @@
+package com.banking.app.repositories;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.banking.app.models.User;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Integer> {
+  Optional<User> getByEmail(String email);
+}
