--- banking-backend/src/main/java/com/banking/exceptions/CannotUpdateUserException.java
@@ -1,14 +1,0 @@
-package com.banking.exceptions;
-
-public class CannotUpdateUserException extends RuntimeException {
-
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-	
-	public CannotUpdateUserException() {
-		super("Cannot update user.");
-	}
-
-}
