--- banking-backend/src/main/java/com/banking/exceptions/ClaimAlreadyProcessedException.java
@@ -1,14 +1,0 @@
-package com.banking.exceptions;
-
-public class ClaimAlreadyProcessedException extends RuntimeException {
-
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-	
-	public ClaimAlreadyProcessedException() {
-		super("Claim already processed.");
-	}
-
-}
