--- banking-backend/src/main/java/com/banking/exceptions/EmailAlreadyExistsException.java
@@ -1,14 +1,0 @@
-package com.banking.exceptions;
-
-public class EmailAlreadyExistsException extends RuntimeException {
-
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-	
-	public EmailAlreadyExistsException() {
-		super("This email address is not available. Please log in or register with a different email address.");
-	}
-
-}
