--- banking-backend/src/main/java/com/banking/exceptions/InvalidCredentialsException.java
@@ -1,14 +1,0 @@
-package com.banking.exceptions;
-
-public class InvalidCredentialsException extends RuntimeException {
-
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-	
-	public InvalidCredentialsException() {
-		super("Incorrect username or password.");
-	}
-
-}
