--- banking-backend/src/main/java/com/banking/exceptions/NotAnEmployeeException.java
@@ -1,14 +1,0 @@
-package com.banking.exceptions;
-
-public class NotAnEmployeeException extends RuntimeException {
-
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-	
-	public NotAnEmployeeException() {
-		super("User does not have the required organizational roles to access this information.");
-	}
-
-}
